refactor(backend): move organization schemas to schemas/organization.py (575 -> 496 lines)
